+++ src/services/http/httpService.js
@@ -1,5 +1,4 @@
-import { CATEGORIES, CATEGORY, DATA, LANGUAGES, SEARCH_ADDRESS } from './endpoints';
-import {getSelectedCheckboxesOfCategory} from '../../components/Map/Map'
+import { CATEGORIES, CATEGORY, DATA, LANGUAGES, LOCATION, LOCATIONS, SEARCH_ADDRESS } from './endpoints';
 
 export const httpService = {
     getCategories: () => fetch(CATEGORIES).then(response => response.json()),
@@ -20,7 +19,18 @@ export const httpService = {
     },
 
     getLocations: async filtersUrlParams => {
-        const response = await fetch(`${DATA}?${filtersUrlParams}`, {
+      const ENDPOINT = window.USE_LAZY_LOADING ? LOCATIONS : DATA;
+      const response = await fetch(`${ENDPOINT}?${filtersUrlParams}`, {
+          method: 'GET',
+          headers: {
+              'Content-Type': 'application/json',
+          },
+      });
+      return response.json();
+    },
+
+    getLocation: async locationId => {
+      const response = await fetch(`${LOCATION}/${locationId}`, {
             method: 'GET',
             headers: {
                 'Content-Type': 'application/json',
@@ -54,4 +64,5 @@ export const httpService = {
 
         return fetch(`${SEARCH_ADDRESS}?${queryString}`).then(response => response.json());
     },
+
 };